New version for modifed Excel file from NIMHANS
This Excel file will require some work to become a good
self-explanatory reference recruitment file. For now we
have added lots of work-arounds and complicted code.

We were under pressure to plan follow-up so we took the
file almost as is.

Once this reference file is cleaned up and becomes a nice
recruitment file, we can adapat our code.